# docs/swagger.yaml
basePath: /
definitions:
  dto.APIErrorResponse:
    properties:
      errors: {}
      message:
        example: Validation error
        type: string
      success:
        example: false
        type: boolean
    type: object
  dto.APIResponse:
    properties:
      data: {}
      message:
        example: Operation successful
        type: string
      success:
        example: true
        type: boolean
    type: object
...
  dto.CreateWasteTypeRequest:
    properties:
      description:
        example: Botol plastik bersih siap daur ulang
        type: string
      is_active:
        example: true
        type: boolean
      name:
        example: Plastik Botol Super
        maxLength: 255
        minLength: 2
        type: string
      points_per_kg:
        example: 550
        minimum: 0
        type: integer
      unit_price_per_kg:
        example: 3500
        minimum: 0
        type: number
    required:
    - name
    - points_per_kg
    - unit_price_per_kg
    type: object
...
  dto.UpdateWasteTypeRequest:
    properties:
      description:
        example: Botol plastik bersih disortir
        type: string
      is_active:
        example: true
        type: boolean
      name:
        example: Plastik Botol Super Edit
        type: string
      points_per_kg:
        example: 600
        type: integer
      unit_price_per_kg:
        example: 3800
        type: number
    type: object
...
  dto.WasteTypeResponse:
    properties:
      created_at:
        type: string
      description:
        example: Botol PET bersih
        type: string
      id:
        example: 1
        type: integer
      is_active:
        example: true
        type: boolean
      name:
        example: Plastik (Botol / Ember / Kresek)
        type: string
      points_per_kg:
        example: 500
        type: integer
      unit_price_per_kg:
        example: 3000
        type: number
      updated_at:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@ecopoints.test
    name: EcoPoints Dev Team
  description: REST API for EcoPoints application built with Go Gin and connected
    to existing Laravel MySQL database.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: EcoPoints Golang REST API
  version: "1.0"
paths:
...
  /api/v1/waste-types:
    get:
      description: Retrieve list of all waste types from database
      produces:
      - application/json
      responses:
        "200":
          description: Waste types retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.WasteTypeResponse'
                  type: array
              type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.APIErrorResponse'
      summary: Get all waste types
      tags:
      - Waste Types
    post:
      consumes:
      - application/json
      description: Create a new waste type (Admin only)
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        type: string
      - description: Waste Type Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateWasteTypeRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Waste type created
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.WasteTypeResponse'
              type: object
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/dto.APIErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.APIErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.APIErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new waste type
      tags:
      - Waste Types
  /api/v1/waste-types/{id}:
    delete:
      description: Delete waste type by ID (Admin only)
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        type: string
      - description: Waste Type ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Waste type deleted
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "400":
          description: Invalid ID
          schema:
            $ref: '#/definitions/dto.APIErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.APIErrorResponse'
        "404":
          description: Waste type not found
          schema:
            $ref: '#/definitions/dto.APIErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.APIErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a waste type
      tags:
      - Waste Types
    get:
      description: Retrieve details of a specific waste type by its ID
      parameters:
      - description: Waste Type ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Waste type retrieved
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.WasteTypeResponse'
              type: object
        "400":
          description: Invalid ID
          schema:
            $ref: '#/definitions/dto.APIErrorResponse'
        "404":
          description: Waste type not found
          schema:
            $ref: '#/definitions/dto.APIErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.APIErrorResponse'
      summary: Get waste type by ID
      tags:
      - Waste Types
    put:
      consumes:
      - application/json
      description: Update waste type by ID (Admin only)
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        type: string
      - description: Waste Type ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated Waste Type Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateWasteTypeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Waste type updated
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.WasteTypeResponse'
              type: object
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/dto.APIErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.APIErrorResponse'
        "404":
          description: Waste type not found
          schema:
            $ref: '#/definitions/dto.APIErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.APIErrorResponse'
      security:
      - BearerAuth: []
      summary: Update an existing waste type
      tags:
      - Waste Types
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'Type "Bearer " followed by your JWT token. Example: "Bearer eyJhbGciOi..."'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"